perf: Complete Decimal to float replacement across all modules
Remove all remaining Decimal imports and usages from:
- All instrument models (bond, repo, interbank, derivatives, etc.)
- Risk calculators (deposit_elasticity, factor_analysis)
- Mock data generator
- Example usage files

This completes the performance optimization to use native float64
instead of Python's Decimal type for all numerical calculations.

Related to previous commit for 50-100x performance improvement.

+++ alm_calculator/risks/liquidity/currency_liquidity_gaps.py
@@ -107,12 +107,12 @@ class CurrencyLiquidityGapCalculator:
         self,
         instruments: List[BaseInstrument],
         risk_params: Dict
-    ) -> Dict[str, Dict[str, Dict[str, Decimal]]]:
+    ) -> Dict[str, Dict[str, Dict[str, float]]]:
         """
         Собирает cash flows по валютам и временным бакетам.
 
         Returns:
-            Dict[currency, Dict[bucket, {'inflow': Decimal, 'outflow': Decimal}]]
+            Dict[currency, Dict[bucket, {'inflow': float, 'outflow': float}]]
         """
         cash_flows = {}
 
@@ -147,7 +147,7 @@ class CurrencyLiquidityGapCalculator:
     def _calculate_gaps_for_currency(
         self,
         currency: str,
-        cash_flows: Dict[str, Dict[str, Decimal]]
+        cash_flows: Dict[str, Dict[str, float]]
     ) -> pd.DataFrame:
         """
         Рассчитывает гэпы ликвидности для одной валюты.
